requirements.txt
# Core ML
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
scikit-learn>=1.4.0
lightgbm>=4.3.0

# RecSys
implicit>=0.7.2

# Generative Modeling
diffusers>=0.27.0
accelerate>=0.27.0

# Vector DB + Retrieval
qdrant-client>=1.8.0
FlagEmbedding>=1.2.0          # bge-large, bge-reranker
rank-bm25>=0.2.2

# LLM + Agent
langchain>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0
langsmith>=0.1.0
openai>=1.30.0
anthropic>=0.25.0
instructor>=1.2.0              # structured outputs

# Multimodal
openai-whisper>=20231117
transformers>=4.40.0
pillow>=10.3.0
TTS>=0.22.0                    # Coqui TTS

# Inference Optimization
bitsandbytes>=0.43.0
triton>=2.3.0

# Eval
ragas>=0.1.7
weave>=0.50.0                  # W&B Weave
wandb>=0.17.0

# Data
pandas>=2.2.0
pyarrow>=16.0.0
requests>=2.31.0

# Dev
pytest>=8.0.0
black>=24.0.0
ruff>=0.4.0
jupyter>=1.0.0
python-dotenv>=1.0.0
